refactor: Split ActionMenuItems to another component
